Add -U UUID option - resolve #1229
Also fixed an error with -u uuidfile where it would not set the test boolean
to true if the UUID was "true".

The -u uuidfile and -U UUID options may not be used with -i answers, -d
or -s seed.

Updated man page for the above changes.

**IMPORTANT NOTE**: you do NOT need to use this update in order to participate
in the IOCCC28!