Partial fix to issue #1208
The -x option to force delete the submission directory and -r
rm option to set path to rm(1) were added to mkiocccentry as part of
issue #1208. The change in order of args is NOT done and will not be
done until AFTER IOCCC28.

Fixed bug where overwrite_answers was always true by default
(mkiocccentry).