Resolved issues #1233 and #1218
Both mkiocccentry and txzchk (the only ones that use other tools) now
search $PATH for the tools first by way of the find_utils() function
(modified a fair bit) and a new util function in jparse called
resolve_path(). The jparse util functions shell_cmd() and pipe_open()
were also improved to resolve paths if no / is in the command name. As
for make: we now search for gmake first as the Makefiles we need are GNU
Makefiles.