change repo release version to 2.4.3 2025-03-10
References to version 2.4.4 and 2.4.5 were done in error.  There has
been no previous releases of this repo since IOCCC28 opened.  As such we
"collapse" these non-releases of this repo into a single release of:

> Release 2.4.3 2025-03-10

When we later announce the first release to this mkiocccentry toolkit
once IOCCC28 opened, we will announce the presence of just one new
repo version.

Please NOTE: This commit, while recommended, is not mandatory for IOCCC28.
             Existing submissions under the previously released mkiocccentry
             toolkit remain valid.