Update comments in mkiocccentry.c
The most important one is for '-i': it suggested we enable yes to
prevent problem of different file sets but that is no longer the case.
The -y is for questions and not prompts (that is what -Y is for, for
better or for worse).

There was also a duplicate line of setting answer_yes to true.

Changed the comments for find_utils() in mkiocccentry.

Obviously this commit is not important other than to clarify things so
people might not think it's a bug that using -y still prompts you for
some things.