improve how the C locale is set
We have expanded the list of C locate environment variables set by shell scripts.

For the code in this repo, we will add a new `set_ioccc_locale()` function to
`soup/location_util.c` to be called in `main()` instead of a direct call
to `(void) setlocale(LC_ALL, "C")`.

The `set_ioccc_locale()` function to `soup/location_util.c` contains a
**programmers apology** that is recommended reading.

The above does **NOT** apply to code under `jparse/`, nor under `dbg/`, or under `dyn_array/`.

Tested the above using the same method as described in
commit 0541e922c794601c5260d76bcc741e71dacad43d.