Add temporary workaround issue #1258
Add temporary workaround for issue #1258. It might be that the real fix
involves an update to a jparse routine but during the IOCCC28 itself
that is not going to be looked at and it is arguable it is not even a
good idea to do that anyway.  Instead for the time being in the
problematic code an extra check is done with another FTS related routine
I had written last year (I guess) in jparse. An XXX comment was added to
chkentry.c but it might be (unsure as this was done quickly) that a fix
in jparse/util.c has to be done instead. That's TBD later on.

Updated CHKENTRY_VERSION to "2.0.4 2025-05-02".