allow submissions to have a larger clock error
Allow a submitter's system clock to be as much as 48 hours in the future before
their submission is rejected as having a bogus timestamp by changing
`MAX_CLOCK_ERROR` from ((42*60)-1) to (48*60*60).

NOTE: The "48 hours" is experimental.  We might change it prior to the
next repo release.

Changed `CHKENTRY_VERSION` from "2.0.4 2025-05-02"
to "2.0.5 2025-05-03".

Changed `MKIOCCCENTRY_REPO_VERSION` from "2.4.7 2025-05-02"
to 2.4.8 2025-05-03".

Please NOTE: This commit, while recommended, is not mandatory for IOCCC28.
             Existing submissions under the previously released mkiocccentry
             toolkit remain valid.