Resolve issue #1207 - ignore SCCS and RCS
It might be worth noting that if one has an ignored directory name as a
file that it appears to not warn about it even though it will (it
appears as well) ignore it. It is unclear to me if this should be
changed or not.

It would seem that both chkentry and txzchk have this problem as well. A
new issue for each will be opened to be resolved before IOCCC29.