Further address issue #1228 and fix a bug
Upon review of what I had in the check_submission_dir() function and
further inspection it is in fact possible and ideal to check that the
submission directory has only files that are in the manifest, assuming
the manifest option was used. This was possible by making sure to set
the manifest paths in the fts struct to NULL (reset_fts() in theory
should do this but it is done explicitly with an assignment) but keeping
the list in the info struct. This way files would be found by the FTS
function and then when a regular file is found (FTS_F) the array in the
info struct is checked. If it a file in the directory is not in the
manifest it is an error.

Fix bug where the function check_submission_dir() did not actually verify if
any ignored files were present in the submission directory. This check
is done in the same manner as the above check except that instead of
making sure the files that are in the directory are in the manifest it
makes sure that the files in the directory are NOT in the ignore list.

A question was posed in the comments that might be worth noting here:

    /*
     * Should we check (here or anywhere else) that an ignored
     * file is not in the manifest? On the one hand it would
     * seem to be a command line error but on the other hand if
     * someone is using the manifest option it probably can be
     * assumed they know what they are doing and given that they
     * have to verify things look okay it should not really
     * matter much (if at all).
     */

If this is deemed not ideal it can be updated.