Address issue #1216 and fix fnmatch(3) bug
Address issue #1216. By using -X file (X for eXclude) one can provide a
file with a list of path to ignore (one path per line). Like with -I
path and -M manifest_file one may have (in the file itself) globs (based
on the rules of fnmatch(3)). The reason a long option is not used is
because GNU and BSD have different behaviours in getopt_long(3) in ways
that might cause problems. This can be looked at later if necessary.

Fix bug with globs in manifest and ignore lists. Now when checking lists
that can have a glob in an item it uses the updated jparse routines that
take a new bool fn.

Updated MKIOCCCENTRY_VERSION to "2.0.12 2025-06-21".
Updated CHKENTRY_VERSION to "2.0.6 2025-06-21".
Updated SOUP_VERSION to "2.0.4 2025-06-21".