Fix bug in append_unique_filename()
Change append_unique_filename() to not abort when the filename already
exists in the array. This is important in case (for example) someone has
a directory named as a forbidden filename. It does not hurt to have the
function return doing nothing and given the error message that was given
it would have been very confusing to someone not familiar with the code.