Remove test case for txzchk for GitHub
An odd problem with GitHub is causing the script to fail in the
workflow. The output in the error file is from txzchk from a text file
so it's not anything to do with tar. The error file was built. The test
works here but not in GitHub so the files were removed in hopes this
solves the problem.