checkpoint draft change to dbg.h to handle booleans
Due to the ANSI C committees continued disregard for backward compatibility,
and due to complications for applications beyond this repo that
include the [dbg.h header file](https://github.com/lcn2/dbg/blob/master/dbg.h),
especially in regards to complications relating to the use of `<stdbool.h>`
(or lack thereof) as well as older C standards and even non-UNIX environments,
that while **BEYOND THE SCOPE** of this repo, are within scope of
the [dbg repo](https://github.com/lcn2/dbg), we propose this draft change
to `dbg/dbg.h`.

As a result of the above, `jparse/util.h` has been changed.  In particular,
that file now depends on `dbg.h` addressing the boolean mess.  Moreover,
it now includes `dbg.h` before `dyn_array.h`, which is what everything
else does in this repo.

TODO: Test the changes make in this checkpoint commit.  If all is well,
      create a `cbool.h` file in the [dbg repo](https://github.com/lcn2/dbg)
      by moving the boolean mess into that file, and let `dbg.h` include
      that new `cbool.h` file.  If all remains well, update this repo
      and issue a PR to the [jparse repo](https://github.com/xexyl/jparse)
      as needed.