improve picky
We will be more picky about how we perform a `make picky`.

Added `${PICKY_OPTIONS}` to specify the picky options
for the `${C_SRC}` and ${H_SRC} files.

Added `${LESS_PICKY_OPTIONS}` to specify the picky options
for the `${LESS_PICKY_CSRC}` and ${LESS_PICKY_HSRC} files.

Added `${SH_PICKY_OPTIONS}` to specify the picky options
for the `${SH_FILES}` files.

Where applicable, added ${BISON_FLEX_PICKY_OPTIONS` to specify the picky options
for the `${BISONFILES}` and ${FLEXFILES} files.

The `make picky` rule tests a single set, if that set is non-empty,
one at a time.

Adjusted the order of `soup/Makefile` adjusting
and organizing the make rules needed to build things.

Ran `make release` to test and perform the above.