add pr repo, improve location comment
We have added the `pr/` tree which an "include" of the
[pr repo](https://github.com/lcn2/pr) code to implement
the "_stdio helper library_" code.

While the `pr/` tree code came from:

- soup/file_util.h
- soup/file_util.c
- jparse/util.h
- jparse/util.c

As of this commit, none of that original code has been
replaced with use of `pr/pr.h` and `pr/libpr.a`.

This new `pr/` tree is NOT yet used by this repo.

TODO: Fill out the man pages, and the `pr_test.c` code in
      [pr repo](https://github.com/lcn2/pr), then update
      this repo to use  `pr/pr.h` and `pr/libpr.a`.