sync with pr repo
From the pr repo git log:

remove minor race condition

As noted in the [Time-of-check
time-of-use](https://github.com/ioccc-src/mkiocccentry/security/code-scanning/55),
a filesystem race condition was possible between the open of a file in
`open_dir_file(char const *dir, char const *file)` between when a file
was checked for readability, and when it was open.  Now the file will
attempt to be opened first, and if it cannot be opened for reading,
a call to `errp()` will occur.  The later check of "must be a readable
file" is performed on the file descriptor, and while less important
after a successful open for reading, will still handle cases with "file"
is not a regular file.

Changed `PR_VERSION` from "1.1.0 2025-09-20"
to "1.1.2 2025-09-22".

Ran `make release` to test the above.