New release of jparse - 2.5.0 2025-10-02
Synced jparse repo (https://github.com/xexyl/jparse) with various fixes
and also what was done with release 2.7.3 2025-09-29 here (some of the
fixes being in that commit but a fair few older ones too). See
jparse/CHANGES.md if you wish to know more.

Also added to the previous release of this repo (yesterday 1 October
2025) the update to soup/rule_count.c - add missing nullptr keyword for
C23+ (instead of what makes more sense - NULL, even if that's a macro)
from @SirWumpus (and perhaps he should get rid of the checks for
trigraphs since the committee decided it would be brilliant to do that
too :( ? :-) ).