improve on pr 1321
Improved `make depend` to not produce needless "`../`" in the dependencies.

Improved how `BITS_IN_BYTE`, `BYTE_VALUES`, and `MAX_BYTE` are defined.

With the exception of printing Unicode Control Point hex values,
regularized to debug, warning and error messages to use lower case hex,
with a leading 0x, and to be consistent with how "the repo" does this.

Ran `make release` to test the above.