add cpath clone from cpath repo
We add `cpath/` as clone from the
[cpath repo](https://github.com/lcn2/cpath.git)
release version **2.0.0 2025-10-16**.

NOTE: This is just a clone addition.  The rest
of the code has not yet been modified to use
the `cpath/` tree.  Moreover the code in the repo
has not yet been updated to use the
version **2.0.0 2025-10-16** of `libcpath.a`:
as of this update, this repo still uses
version **1.0.2 2025-10-12**.

Ran `make release` to test the above.

TODO: Integrate the code in this repo to use
      version **2.0.0 2025-10-16** of `libcpath.a`.