checkpoint, add calls to processing the walk
Added calls for `fts_walk()`, `sort_walk_istat()`, `chk_walk()`,
and `scan_topdir()` to `mkiocccentry.c`.

The `chk_walk()` just writes error messages to `stderr`.
The `mkiocccentry()` tool may or may NOT wish to pass `NULL`
to the stream and manually process the result of the walk.
On the other hand, the current use of `chk_walk()` function
may be good enough.

The calls too `err(3)` on the failure of `fts_walk()`, and on the the
failure of `chk_walk()` need to be considered.  The mkiocccentry()`
tool may or may NOT wish to change how these errors are handled.  On the
other hand, the current use of `err(3)` may be good enough.

The above adds to what commit 8dec57012db13276bfd6cc24dabfd16b883383fa
added in terms of the initialization of the `struct walk_stat wstat`
and the `-I path` arg processing.

**NOTE**: The old code for functions such as `scan_topdir()`,
which is an obsolete function, have not been modified.
This is just a checkpoint.

Ran `make release` to test the above.