add more feather testing to txzchk_test.sh
Added some additional tests for `test_ioccc/txzchk_test.sh` to process
in order to further improve the test case coverage for `txzchk(1)`.

Ran `make release` to test the above.

TODO:

In regards to what `test_ioccc/txzchk_test.sh` tests:

Look over the new tests (both good and bad) to be sure that
the test results are OK.

From what we can see, most of the new tests seem OK except for:

Fix: `submit.12345678-1234-4321-abcd-1234567890ab-8.1761934327.txt.err`

The output should **NOT** be empty due to a negative file length.

Review: `submit.12345678-1234-4321-abcd-1234567890ab-8.1761934327.txt.err`

Is the "_Warning: string_to_intmax: error converting string_"
the best way to handle "super-bogus" lengths?  Maybe it is OK?

Fix: `submit.12345678-1234-4321-abcd-1234567890ab-1.1761934316.txt.err`

The setuid shell script was not detected.
The setgid shell script was not detected.

Fix: `submit.12345678-1234-4321-abcd-1234567890ab-0.1761939801.txt.err`

The quasi-macOS style `-r--r--r--@` in the 1st filed of a tarball listing
should be flagged as an error.  Yes, this is a contrived test case in that
this is NOT how such a tarball listing would be presented.  Nevertheless, this
suggests the parsing of a tarball listing may need to made more robust.

Fix: `submit.12345678-1234-4321-abcd-1234567890ab-0.1761940107.txt.err`

Adding a leading space to the 1st filed of a tarball listing
should be flagged as an error.  Yes, this is a contrived test case in that
this is NOT how such a tarball listing would be presented.  Nevertheless, this
suggests the parsing of a tarball listing may need to made more robust.

Fix: `submit.12345678-1234-4321-abcd-1234567890ab-0.1761940499.txt.err`

Adding a trailing "0" to the 1st filed of a tarball listing
should be flagged as an error.  Yes, this is a contrived test case in that
this is NOT how such a tarball listing would be presented.  Nevertheless, this
suggests the parsing of a tarball listing may need to made more robust.

Fix: `submit.12345678-1234-4321-abcd-1234567890ab-0.1761941033.bin.err`

The contents is random data.  Nevertheless these error messages
indicate that there may be problems with processing random garbage
that may spell trouble for the `txzchk(1)` tool.  Nevertheless, this
suggest the parsing of a tarball listing may need to made more robust.

```
Warning: txzchk: found NUL before end of line: errno[0]: Undefined error: 0
Warning: txzchk: found NUL before end of line: errno[0]: Undefined error: 0
Warning: readline: line does not end in newline: <<random garbage omitted>>
ERROR[39]: parse_txz_line: txzchk: test_ioccc/test_txzchk/bad/submit.12345678-1234-4321-abcd-1234567890ab-0.1761941033.txt: NULL pointer encountered trying to parse line
```