add even more feather testing to txzchk_test.sh
Added some additional tests for `test_ioccc/txzchk_test.sh` to process
in order to further improve the test case coverage for `txzchk(1)`.

Ran `make release` to test the above.

TODO:

In regards to what `test_ioccc/txzchk_test.sh` tests:

Look over the additional bad tests to be sure that the test results are OK.

From what we can see, most of the additional bad tests seem OK except for:

Fix: `submit.12345678-1234-4321-abcd-1234567890ab-0.1761942803.txt.err`

The "file is not executable and set-user-ID mode is set" was not detected.

Fix: `submit.12345678-1234-4321-abcd-1234567890ab-0.1761942974.txt.err`

The "file is not executable and set-group-ID mode is set" was not detected.

Fix: `submit.12345678-1234-4321-abcd-1234567890ab-0.1761943429.txt.err`

The "file has an ACL" was not detected.

Fix: `submit.12345678-1234-4321-abcd-1234567890ab-0.1761943566.txt.err`

A file with mode 0000 was not detected.

Fix: `submit.12345678-1234-4321-abcd-1234567890ab-9.1761943766.txt.err`

A directory with mode 0000 was not detected.

**REQUEST**: The `txzchk(1)` tool should enforce the permissions that
`chkentry -S` requires.  In particular directories should be `drwxr-xr-x`,
files that do **NOT** end in `.sh` should be `-r--r--r--`,
and shell script files that end in `.sh` should be `-r-xr-xr-x`.