improve pr 1336
Allow for empty lines and lines beginning with "#" (hash) are ignored
in manifest files to allow for comments and blank lines.

Allow for empty lines and lines beginning with "#" (hash) are ignored
in ignore files to allow for comments and blank lines.

Changed `warn(3)` calls related to duplicates into `dbg(3)` calls in `read_manifest()`.

Fixed a few whitespace issues in `mkiocccentry.c`.
Fixed a few whitespace issues in `soup/file_util.c`.

Ran `make release` to test the above.