Fix semantics errors from commit 63a719a
It should be noted that now that mkiocccentry uses the table code and
simply skips over the files that are too long in name or for other
reasons, the mkiocccentry_test.sh script fails as the bad tests no
longer fail (or some don't if not all). A new issue will be opened to
discuss this issue. The chksubmit_test.sh has an issue too or so it
seems; testing it manually here it passed but it seemed like it failed
in make test. This will be mentioned in the new issue that will be
opened but the semantics errors should be fixed at least.