Fix issue #1355
Fix make install to NOT install repo specific man pages and make
uninstall to remove some man pages (and possibly scripts) that were
erroneously installed (and also some that were not removed). This was
done with the jparse/ repo too (synced here). It would appear the other
repos did not have this problem.

A cosmetic fix was made in the top level Makefile as well (in a
comment).