rename MAX_CLOCK_ERROR to FUTURE_CLOCK_SKEW_LIMIT
The `FUTURE_CLOCK_SKEW_LIMIT` is not a real limit, so it does not need
to be in `soup/limit_ioccc.h`, nor should the value be in the same
"scope" as other "_MAX_" symbols in `soup/limit_ioccc.h`.

Ran `make release` to test the above.

p.s. Commit from a remote site with reliable Internet. _sigh_